chore: changes PR feedback
